Implement `freezed` for parsing API objects
This also means using the JSON deserializer.
Not all objects are implemented yet, but enough are that the application retains its functionality.
Not all fields have been tested, and as the parser is much more strict when it comes to types, it may require more work to ensure it works with arbitrary GraphQL queries.