Upgrade Flutter and many other dependencies
I needed to create a new release and was unsure of the state of the
project. These changes were made by:
* Upgrading Flutter to the latest version
* Upgrading Android Studio to the latest version
* Upgrading any dependencies that weren't pinned

Then working through every problem until the project built again. This
includes changing the Gradle version and switching some Flutter
dependencies.

A few bugs are present, but the project builds and seems to work on the
main screen at least. Future work can begin from this "working" state.